# [RFC](Don't merge) Faster platform-specific compiles