--- a/application/application.d.ts
+++ b/application/application.d.ts
@@ -4,6 +4,7 @@
 declare module "application" {
     import cssSelector = require("ui/styling/css-selector");
     import observable = require("data/observable");
+    import native_api = require("native-api");
 
     /**
      * An extended JavaScript Error which will have the nativeError property initialized in case the error is caused by executing platform-specific code.
@@ -209,7 +210,7 @@ declare module "application" {
         /**
          * The activity.
          */
-        activity: android.app.Activity;
+        activity: native_api.android.app.Activity;
 
         /**
          * The name of the event.
@@ -229,7 +230,7 @@ declare module "application" {
         /**
          * The bundle.
          */
-        bundle: android.os.Bundle;
+        bundle: native_api.android.os.Bundle;
     }
 
     /**
@@ -249,7 +250,7 @@ declare module "application" {
         /**
          * The intent.
          */
-        intent: android.content.Intent;
+        intent: native_api.android.content.Intent;
     }
 
     /**
@@ -269,27 +270,27 @@ declare module "application" {
         /**
          * The [android Application](http://developer.android.com/reference/android/app/Application.html) object instance provided to the init of the module.
          */
-        nativeApp: android.app.Application;
+        nativeApp: native_api.android.app.Application;
 
         /**
          * The application's [android Context](http://developer.android.com/reference/android/content/Context.html) object instance.
          */
-        context: android.content.Context;
+        context: native_api.android.content.Context;
 
         /**
          * The currently active (loaded) [android Activity](http://developer.android.com/reference/android/app/Activity.html). This property is automatically updated upon Activity events.
          */
-        foregroundActivity: android.app.Activity;
+        foregroundActivity: native_api.android.app.Activity;
 
         /**
          * The currently active (loaded) Context. This is typically the top-level Activity that is just created.
          */
-        currentContext: android.content.Context;
+        currentContext: native_api.android.content.Context;
 
         /**
          * The main (start) Activity for the application.
          */
-        startActivity: android.app.Activity;
+        startActivity: native_api.android.app.Activity;
 
         /**
          * The name of the application package.
@@ -301,47 +302,47 @@ declare module "application" {
          * @param intent - Native (android) intent used to create the activity.
          * Returns com.tns.NativeScriptActivity.extend implementation.
          */
-        getActivity(intent: android.content.Intent): any;
+        getActivity(intent: native_api.android.content.Intent): any;
 
         /**
          * [Deprecated. Please use the respective event instead.] Direct handler of the [onActivityCreated method](http://developer.android.com/reference/android/app/Application.ActivityLifecycleCallbacks.html).
          */
-        onActivityCreated: (activity: android.app.Activity, bundle: android.os.Bundle) => void;
+        onActivityCreated: (activity: native_api.android.app.Activity, bundle: native_api.android.os.Bundle) => void;
 
         /**
          * [Deprecated. Please use the respective event instead.] Direct handler of the [onActivityDestroyed method](http://developer.android.com/reference/android/app/Application.ActivityLifecycleCallbacks.html).
          */
-        onActivityDestroyed: (activity: android.app.Activity) => void;
+        onActivityDestroyed: (activity: native_api.android.app.Activity) => void;
 
         /**
          * [Deprecated. Please use the respective event instead.] Direct handler of the [onActivityDestroyed method](http://developer.android.com/reference/android/app/Application.ActivityLifecycleCallbacks.html).
          */
-        onActivityStarted: (activity: android.app.Activity) => void;
+        onActivityStarted: (activity: native_api.android.app.Activity) => void;
 
         /**
          * [Deprecated. Please use the respective event instead.] Direct handler of the [onActivityPaused method](http://developer.android.com/reference/android/app/Application.ActivityLifecycleCallbacks.html).
          */
-        onActivityPaused: (activity: android.app.Activity) => void;
+        onActivityPaused: (activity: native_api.android.app.Activity) => void;
 
         /**
          * [Deprecated. Please use the respective event instead.] Direct handler of the [onActivityResumed method](http://developer.android.com/reference/android/app/Application.ActivityLifecycleCallbacks.html).
          */
-        onActivityResumed: (activity: android.app.Activity) => void;
+        onActivityResumed: (activity: native_api.android.app.Activity) => void;
 
         /**
          * [Deprecated. Please use the respective event instead.] Direct handler of the [onActivityStopped method](http://developer.android.com/reference/android/app/Application.ActivityLifecycleCallbacks.html).
          */
-        onActivityStopped: (activity: android.app.Activity) => void;
+        onActivityStopped: (activity: native_api.android.app.Activity) => void;
 
         /**
          * [Deprecated. Please use the respective event instead.] Direct handler of the [onActivitySaveInstanceState method](http://developer.android.com/reference/android/app/Application.ActivityLifecycleCallbacks.html).
          */
-        onSaveActivityState: (activity: android.app.Activity, bundle: android.os.Bundle) => void;
+        onSaveActivityState: (activity: native_api.android.app.Activity, bundle: native_api.android.os.Bundle) => void;
 
         /**
          * [Deprecated. Please use the respective event instead.] Direct handler of the onActivityResult method.
          */
-        onActivityResult: (requestCode: number, resultCode: number, data: android.content.Intent) => void;
+        onActivityResult: (requestCode: number, resultCode: number, data: native_api.android.content.Intent) => void;
             
         /**
          * A basic method signature to hook an event listener (shortcut alias to the addEventListener method).
@@ -447,7 +448,7 @@ declare module "application" {
          * @param intentFilter A string containing the intent filter.
          * @param onReceiveCallback A callback function that will be called each time the receiver receives a broadcast.
          */
-        registerBroadcastReceiver(intentFilter: string, onReceiveCallback: (context: android.content.Context, intent: android.content.Intent) => void): void;
+        registerBroadcastReceiver(intentFilter: string, onReceiveCallback: (context: native_api.android.content.Context, intent: native_api.android.content.Intent) => void): void;
 
         /**
          * Unregister a previously registered BroadcastReceiver. 
@@ -466,12 +467,12 @@ declare module "application" {
         /**
          * The root view controller for the application.
          */
-        rootController: UIViewController;
+        rootController: native_api.UIViewController;
 
         /**
          * The [UIApplication](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UIApplication_Class/index.html) object instance provided to the init of the module.
          */
-        nativeApp: UIApplication;
+        nativeApp: native_api.UIApplication;
 
         /**
          * Adds an observer to the default notification center for the specified notification.
@@ -479,7 +480,7 @@ declare module "application" {
          * @param notificationName A string containing the name of the notification.
          * @param onReceiveCallback A callback function that will be called each time the observer receives a notification.
          */
-        addNotificationObserver(notificationName: string, onReceiveCallback: (notification: NSNotification) => void): void;
+        addNotificationObserver(notificationName: string, onReceiveCallback: (notification: native_api.NSNotification) => void): void;
 
         /**
          * Removes the observer for the specified notification from the default notification center.