--- a/apps/animations/main-page.ts
+++ b/apps/animations/main-page.ts
@@ -5,6 +5,7 @@ import abs = require("ui/layouts/absolute-layout");
 import animationModule = require("ui/animation");
 import colorModule = require("color");
 import model = require("./model");
+import native_api = require("native-api");
 
 var vm = new model.ViewModel();
 
@@ -27,7 +28,7 @@ export function pageLoaded(args: observable.EventData) {
 
 export function onSlideOut(args: observable.EventData) {
     console.log("onSlideOut");
-    var curve = page.android ? new android.view.animation.AccelerateInterpolator(1) : UIViewAnimationCurve.UIViewAnimationCurveEaseIn;
+    var curve = page.android ? new native_api.android.view.animation.AccelerateInterpolator(1) : native_api.UIViewAnimationCurve.UIViewAnimationCurveEaseIn;
 
     var buttonAnimations = [
         { target: button1, translate: { x: -240, y: 0 }, scale: { x: 0.5, y: 0.5 }, opacity: 0, duration: vm.duration, delay: 0, iterations: vm.iterations, curve: curve },
@@ -45,7 +46,7 @@ export function onSlideOut(args: observable.EventData) {
 
 export function onSlideIn(args: observable.EventData) {
     console.log("onSlideIn");
-    var curve = page.android ? new android.view.animation.DecelerateInterpolator(1) : UIViewAnimationCurve.UIViewAnimationCurveEaseOut;
+    var curve = page.android ? new native_api.android.view.animation.DecelerateInterpolator(1) : native_api.UIViewAnimationCurve.UIViewAnimationCurveEaseOut;
 
     panelAnimation = panel.createAnimation({ opacity: 1, scale: { x: 1, y: 1 }, rotate: 0, backgroundColor: new colorModule.Color("yellow"), duration: vm.duration, iterations: vm.iterations });
 
@@ -86,7 +87,7 @@ export function onSingle(args: observable.EventData) {
         duration: vm.duration,
         delay: 0,
         iterations: vm.iterations,
-        curve: button1.ios ? UIViewAnimationCurve.UIViewAnimationCurveEaseIn : new android.view.animation.AccelerateInterpolator(1),
+        curve: button1.ios ? native_api.UIViewAnimationCurve.UIViewAnimationCurveEaseIn : new native_api.android.view.animation.AccelerateInterpolator(1),
     })
         .then(() => console.log("Animation finished"))
         .catch((e) => console.log(e.message));