--- a/apps/notifications-demo/app.ts
+++ b/apps/notifications-demo/app.ts
@@ -1,13 +1,14 @@
 ﻿import application = require("application");
+import native_api = require("native-api");
 
 application.mainModule = "main-page";
 
 application.on(application.exitEvent, () => {
     if (application.android) {
-        application.android.unregisterBroadcastReceiver(android.content.Intent.ACTION_BATTERY_CHANGED);
+        application.android.unregisterBroadcastReceiver(native_api.android.content.Intent.ACTION_BATTERY_CHANGED);
     }
     else {
-        application.ios.removeNotificationObserver(UIDeviceBatteryLevelDidChangeNotification);
+        application.ios.removeNotificationObserver(native_api.UIDeviceBatteryLevelDidChangeNotification);
     }
 });