--- a/apps/notifications-demo/main-page.ts
+++ b/apps/notifications-demo/main-page.ts
@@ -2,6 +2,7 @@
 import observable = require("data/observable");
 import pages = require("ui/page");
 import labelModule = require("ui/label");
+import native_api = require("native-api");
 
 var batteryLabel: labelModule.Label;
 var registered = false;
@@ -14,10 +15,10 @@ export function onPageLoaded(args: observable.EventData) {
     }
 
     if (application.android) {
-        application.android.registerBroadcastReceiver(android.content.Intent.ACTION_BATTERY_CHANGED,
-            function onReceiveCallback(context: android.content.Context, intent: android.content.Intent) {
-                var level = intent.getIntExtra(android.os.BatteryManager.EXTRA_LEVEL, -1);
-                var scale = intent.getIntExtra(android.os.BatteryManager.EXTRA_SCALE, -1);
+        application.android.registerBroadcastReceiver(native_api.android.content.Intent.ACTION_BATTERY_CHANGED,
+            function onReceiveCallback(context: native_api.android.content.Context, intent: native_api.android.content.Intent) {
+                var level = intent.getIntExtra(native_api.android.os.BatteryManager.EXTRA_LEVEL, -1);
+                var scale = intent.getIntExtra(native_api.android.os.BatteryManager.EXTRA_SCALE, -1);
                 var percent = (level / scale) * 100.0;
                 var message = "Battery: " + percent + "%";
                 console.log(message);
@@ -25,15 +26,15 @@ export function onPageLoaded(args: observable.EventData) {
             });
     }
     else {
-        var onReceiveCallback = function onReceiveCallback(notification: NSNotification) {
-            var percent = UIDevice.currentDevice().batteryLevel * 100;
+        var onReceiveCallback = function onReceiveCallback(notification: native_api.NSNotification) {
+            var percent = native_api.UIDevice.currentDevice().batteryLevel * 100;
             var message = "Battery: " + percent + "%";
             console.log(message);
             batteryLabel.text = message;
         }
-        UIDevice.currentDevice().batteryMonitoringEnabled = true;
+        native_api.UIDevice.currentDevice().batteryMonitoringEnabled = true;
         onReceiveCallback(null);
-        application.ios.addNotificationObserver(UIDeviceBatteryLevelDidChangeNotification, onReceiveCallback);
+        application.ios.addNotificationObserver(native_api.UIDeviceBatteryLevelDidChangeNotification, onReceiveCallback);
     }
     registered = true;
 }