--- a/global_alias.ios.ts
+++ b/global_alias.ios.ts
@@ -1,0 +1,2 @@
+declare module android {};
+declare module java {};