--- a/global_alias.ts
+++ b/global_alias.ts
@@ -1,0 +1,2 @@
+export import ad = android;
+export import aj = java;