--- a/location/location.d.ts
+++ b/location/location.d.ts
@@ -2,6 +2,7 @@
  * Allows you to use the geolocation of the device.
  */
 declare module "location" {
+    import native_api = require("native-api");
     // For future usage
     //class LocationRegion {
     //    public latitude: number;
@@ -57,12 +58,12 @@ declare module "location" {
        /**
         * The android-specific [location](http://developer.android.com/reference/android/location/Location.html) object.
         */
-        android: android.location.Location;
+        android: native_api.android.location.Location;
 
        /**
         * The ios-specific [CLLocation](https://developer.apple.com/library/ios/documentation/CoreLocation/Reference/CLLocation_Class/) object.
         */
-        ios: CLLocation;
+        ios: native_api.CLLocation;
     }
 
    /**
@@ -162,7 +163,7 @@ declare module "location" {
 		/**
 		 * The android-specific location manager [LocationManager](http://developer.android.com/reference/android/location/LocationManager.html)
 		 */
-		manager: android.location.LocationManager;
+		manager: native_api.android.location.LocationManager;
 
 		/**
 		 * The minimum time interval between subsequent location updates, in milliseconds.
@@ -186,6 +187,6 @@ declare module "location" {
 		/**
 		 * The ios-specific location manager [CLLocationManager](https://developer.apple.com/library/ios/documentation/CoreLocation/Reference/CLLocationManager_Class/)
 		 */
-		manager: CLLocationManager;
+		manager: native_api.CLLocationManager;
 	}
 }