--- a/ui/core/view.android.d.ts
+++ b/ui/core/view.android.d.ts
@@ -1,0 +1,7 @@
+declare module "ui/core/view" {
+    //@private
+    export class NativeViewGroup extends android.view.ViewGroup {
+        constructor(context: android.content.Context);
+        public setOwner(view: View);
+    }
+}