Our modules compilation is very very slow. On my machine it takes about 36 seconds. That's just the typescript compilation. It turns out the cause for that are the huge type definitions for the iOS and Android platforms. My attempted solution:

- Creates compilation targets for just Android or just iOS. You should be able to build for just one platform since most of the time you are testing on one of them. When switching platforms, you will simply switch platform configs.
- Aliases unused types to `any`, to avoid forking code or preprocessing.
- All native type access happens through the "native-api" module. It has several implementations that get picked up by the gruntfile when building for a specific configuration.

    - native-api-android.android.d.ts - alias the real android types
    - native-api-android.ios.d.ts - alias the android types to `any`
    - native-api-ios.ios.d.ts - alias the real ios types
    - native-api-ios.android.d.ts - alias the ios types to `any`

Production builds should use native-api-android.android.d.ts and
native-api.ios.ios.d.ts.

# Results:
Full compile: **36 seconds**
Android compile: **23 seconds**
iOS compile: **9 seconds**

# Next steps

Come up with a build target that doesn't pull in any of the platform definitions (just types aliased to <any>) and get to a lightning fast development version TS compile. I can work on that one once this gets tested and gets merged to master.