@hdeshev 2015-07-30
What I think needs to get done before this gets merged to master:

- Consider the public API impact. I hope that one will be pretty minimal since android/ios types should still be perfectly usable.
- Do we need to update any of the docs?
- Get the builds working. Make sure aliased types resolve to the correct ones when running on their platforms.
- Test in Visual Studio. Make sure we don't pollute IntelliSense suggestions with weird types (more than necessary)
- Set up build options and VS configs for the rest of the team.
- Test with a user project (both JavaScript and TypeScript Hello world) and make sure the CLI tool doesn't break with this setup.