@hdeshev 2015-07-30
CC-ing @atanasovg and @ligaz