@ligaz 2015-08-02
I like the build time improvements but the code changes might look strange for the untrained eye. Have you considered splitting the platform d.ts files into smaller ones and require just them? Will that improve the build times or all declarations will be loaded again?