@hdeshev 2015-08-03
@ligaz Splitting the platform files is probably the best solution. I was reluctant to go there because I assumed we'd have to compile all the split files and compile the same (huge) amount of code, but maybe I was wrong. We could exclude all platform.d.ts files from the grunt/tsconfig compile fileset and explicitly `<reference path='..some-file-we-actually-use' />` only a subset of those files.

Another approach that could work is building a tool that will go over the huge platform files and produce a subset of the definitions containing just the types our code actually uses.

The first solution looks like the better alternative since it would not only help our internal development process, but it would also be usable in customer solutions. It does have some downsides though -- for example I don't know how to split the `ios.d.ts` definitions since they are not organized in modules (just a huge list of global types/vars).