@ligaz 2015-08-03
I believe the iOS team already had an idea how to split those. @fealebenpae willing to comment?