@fealebenpae 2015-08-03
I wanted to split the .d.ts output for iOS APIs into chunks that map onto Clang modules so that we get `Foundation.d.ts`, `UIKit.d.ts`, `CoreGraphics.d.ts`, etc. The initial implementation produced external module declarations that had to be `require`d but we postponed that until the runtime supported it. It would be trivial however to change the output to use `///` references instead if that would have a positive impact on the compilation process.