@PanayotCankov 2015-08-03
I would like to check the build too but I don't have time at the moment. sample-Xpensity builds TypeScript in 8 sec. and have the single big iOS d.ts.