@hdeshev 2015-08-03
I had a chat with @atanasovg, and we decided we'd better concentrate our effort on splitting the d.ts files and avoid type aliasing tricks (it would be very confusing to customers). Closing this PR.